create uri handler
